Agregada la parte de sesion en Eventos, Bares y Promociones. Ademas tambien separé la parte de admin a la de bares

+++ app/Http/Controllers/EventController.php
@@ -4,7 +4,9 @@ namespace App\Http\Controllers;
 
 use App\Event;
 use App\City;
+use App\Bar;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Image;
 use File;
 use DB;
@@ -13,7 +15,8 @@ class EventController extends Controller
 {
     public function index()
     {
-        $events = Event::paginate(20);
+        $bar = Bar::findOrFail(Auth::id());
+        $events = $bar->events;
         return view('events.index',compact('events'));
     }
 
@@ -45,6 +48,7 @@ class EventController extends Controller
         $event->lng = $request->lng;
         $event->description = $request->description;
         $event->city_id = $request->city_id;
+        $event->bar_id = Auth::id();
         $event->enabled = $request->enabled;
         $event->date = $request->date;
     
@@ -66,14 +70,22 @@ class EventController extends Controller
     public function show($id)
     {
         $event = Event::findOrFail($id);
-        return view('events.show', compact('event'));
+        if ($event->bar->id == Auth::id())
+            return view('events.show', compact('event'));
+        else
+            return back()->with('error', 'Este evento no le pertenece');
     }
 
    public function edit($id)
     {
         $event = Event::findOrFail($id);
-        $cities = City::pluck('name','id');
-        return view('events.edit', compact('event','cities'));
+        if ($event->bar->id == Auth::id())
+        {
+            $cities = City::pluck('name','id');
+            return view('events.edit', compact('event','cities'));
+        }
+        else
+            return back()->with('error', 'Este evento no le pertenece');
     }
 
     public function update(Request $request, $id)
@@ -89,48 +101,58 @@ class EventController extends Controller
             'enabled' => 'required|boolean',
             'date' => 'required|date',
         ]);
-        
+
         $event = Event::findOrFail($id);
-        $event->title = $request->title;
-        $event->address = $request->address;
-        $event->lat = $request->lat;
-        $event->lng = $request->lng;
-        $event->description = $request->description;
-        $event->city_id = $request->city_id;
-        $event->enabled = $request->enabled;
-        $event->date = $request->date;
-        //Save and get Id
-        $event->save();
-
-        //Image
-        if($request->hasFile('image')) {
-            //If has image, delete it before update
-            $image = public_path('storage/images/events/').$event->image;
-            if(File::exists($image) && $event->image != 'default.jpg') {
-                File::delete($image);
-            }
-            $image = $request->file('image');
-            $event->image = $event->id.time().'.'.$image->getClientOriginalExtension();
-            
-            //Store image
-            Image::make($image)->fit(250, 250)->save(public_path('storage/images/events/') . $event->image );
-            //Save Image info with ID
+        if ($event->bar->id == Auth::id())
+        {
+            $event->title = $request->title;
+            $event->address = $request->address;
+            $event->lat = $request->lat;
+            $event->lng = $request->lng;
+            $event->description = $request->description;
+            $event->city_id = $request->city_id;
+            $event->enabled = $request->enabled;
+            $event->date = $request->date;
+            //Save and get Id
             $event->save();
-        }
 
-        return redirect()->route('events.show', ['id' => $event->id])->with('success', 'Evento editado correctamente!');
+            //Image
+            if($request->hasFile('image')) {
+                //If has image, delete it before update
+                $image = public_path('storage/images/events/').$event->image;
+                if(File::exists($image) && $event->image != 'default.jpg') {
+                    File::delete($image);
+                }
+                $image = $request->file('image');
+                $event->image = $event->id.time().'.'.$image->getClientOriginalExtension();
+                
+                //Store image
+                Image::make($image)->fit(250, 250)->save(public_path('storage/images/events/') . $event->image );
+                //Save Image info with ID
+                $event->save();
+            }
+
+            return redirect()->route('events.show', ['id' => $event->id])->with('success', 'Evento editado correctamente!');
+        }
+        else
+            return back()->with('error', 'Este evento no le pertenece');
     }
 
     public function destroy($id)
     {
         $event = Event::findOrFail($id);
-        //Remove Image
-        $image = public_path('storage/images/events/').$event->image;
-        if(File::exists($image)) {
-            File::delete($image);
+        if ($event->bar->id == Auth::id())
+        {
+            //Remove Image
+            $image = public_path('storage/images/events/').$event->image;
+            if(File::exists($image)) {
+                File::delete($image);
+            }
+            $event->delete();
+            
+            return redirect()->route('events.index')->with('success', 'Evento eliminado correctamente!');
         }
-        $event->delete();
-        
-        return redirect()->route('events.index')->with('success', 'Evento eliminado correctamente!');
+        else
+            return back()->with('error', 'Este evento no le pertenece');
     }
 }
